Refactor outdated warning
